parser color integration
The parser now fully supports colors along with type colors. I didnt test everything yet but the base is well done and im proud of the result so far.

This needs a LOT of code cleanup and optimization, its at a point where i dont fully understand the works of everything anymore and the readability is also increasingly getting worst.

I also need to think about the functionality of the TypeThemes and structure of it as i dont know if im happy with that yet.