Back from break
It has been a month since i last worked on this project and got a bit detached with whatever progress i was making before the break, so this commit is to protect the project from misunderstanding.

The biggest thing that now needs to be fixed is that the color handler also applies color to the color flags which is a problem as it is somewhat of a recursion.
Planning to fix this by first checking if its a color flag of any kind and then excluding that part of the string. Might want to reinforce this by surrounding it with a unique symbol or combination of symbols.