starting over again
the apply functions is starting from nothing once again. this time i really need to write something comprehensible for scalability. My main plan is to write a string constructor that is first just an array of all parts that are either default theme or anything other then default, this array is the whole string split into parts where the theme changes.