partial sub-typeThemes
it is not possible to only partially fill the sub fields of a type, like object. I made all the fields in typethemes optional and adjusted the parser _getTypeTheme function to allow this.
This seems pretty risky so I like to keep an eye on this and see if anything breaks.
#28